Alterações em algumas classes controller e models existentes

+++ src/models/Desconto.java
@@ -2,5 +2,5 @@ package models;
 
 public interface Desconto {
 
-    double calcularDesconto(double valorTotal, int duracaoAluguel);
+    double calcularDesconto(double valorTotal, long duracaoAluguel);
 }
